pyproject.toml
[build-system]
requires      = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name        = "iris_timelineexport_module"
version     = "1.0.0"
description = "IRIS Module – DFIR-Report-style timeline diagram export"
readme      = "README.md"
license     = { file = "LICENSE" }
authors     = [{ name = "baseVISION AG" }]
requires-python = ">=3.9"
# Pillow and pytz are required at runtime; both are present in IRIS containers
# but declared here so the package is installable in other environments too.
dependencies = [
    "Pillow>=9.0",
    "pytz",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where   = ["."]
include = ["iris_timelineexport_module*"]
